+++ terminalManager.ts
@@ -6,7 +6,6 @@ export class TerminalManager {
     private lastSocket;
     private lastTid;
     private lastCreatedTerminal;
-    private lastTerminalIsShared;
     private vfsid;
 
     constructor(
@@ -22,8 +21,7 @@ export class TerminalManager {
                 "terminal": this.lastCreatedTerminal,
                 "pid": parseInt(pty["pid"]),
                 "tid": this.lastTid,
-                "socket": this.lastSocket,
-                "shared": this.lastTerminalIsShared
+                "socket": this.lastSocket
             };
             
             this.eventEmitter.emit('send_ch4_message',
@@ -75,11 +73,14 @@ export class TerminalManager {
         let terminalPath = this.getTerminalPath();
 
         this.vfsid = vfsid;
-        this.lastTerminalIsShared = shared;
 
         if (terminalPath == null) {
             vscode.window.showErrorMessage("Unsupported platform for terminal");
             return;
+        }
+
+        if (shared) {
+            vscode.window.showInformationMessage("Shared terminal requires Visual Studio Code 1.26.0 or greater");
         }
     
         const server = net.createServer((socket) => {
@@ -106,12 +107,6 @@ export class TerminalManager {
                 this.eventEmitter.emit('send_ch4_message',
                     ["write", correctTerminalId.toString(), data.toString()]
                 ); // TODO: Fix
-
-                if (shared) {
-                    this.eventEmitter.emit('send_ch4_message',
-                        ["call","collab","send",[this.vfsid,{"type":"GENERIC_BROADCAST","data":{"exttype":"terminal_udata","tid":correctTerminalId.toString(),"data":data.toString()}}]]
-                    );
-                }
             });
     
             this.lastTid = Math.floor(900*Math.random()) + 100;
@@ -119,12 +114,6 @@ export class TerminalManager {
             this.eventEmitter.emit('send_ch4_message',
                 ["tmux","",{"cwd":"/home/ec2-user/environment","cols":125,"rows":33,"name":"xterm-color","base":"/home/ec2-user/.c9","attach":false,"session":"cloud9_terminal_" + this.lastTid,"output":false,"terminal":true,"detachOthers":true,"defaultEditor":false,"encoding":"utf8","command":"bash -l"},{"$":90}]
             );
-
-            if (shared) {
-                this.eventEmitter.emit('send_ch4_message',
-                    ["call","collab","send",[this.vfsid,{"type":"GENERIC_BROADCAST","data":{"exttype":"terminal_create","tid":this.lastTid}}]]
-                );
-            }
     
             console.log("init'd remote terminal");
         }).on('error', (err) => {
@@ -138,10 +127,6 @@ export class TerminalManager {
         }, () => {
             let addr = server.address();
             let title = "Cloud9 Terminal";
-
-            if (shared) {
-                title = "Cloud9 Terminal (shared)";
-            }
 
             console.log('opened server on', addr);
 
@@ -174,23 +159,11 @@ export class TerminalManager {
 
     closeTerminal(terminal) {
         terminal['socket'].destroy();
-
-        if (terminal['shared']) {
-            this.eventEmitter.emit('send_ch4_message',
-                ["call","collab","send",[this.vfsid,{"type":"GENERIC_BROADCAST","data":{"exttype":"terminal_destroy","tid":terminal['tid']}}]]
-            );
-        }
     }
 
     emitTerminalData(terminal, data) {
         if (typeof data == "string") {
             terminal['socket'].write(data);
-        }
-
-        if (terminal['shared']) {
-            this.eventEmitter.emit('send_ch4_message',
-                ["call","collab","send",[this.vfsid,{"type":"GENERIC_BROADCAST","data":{"exttype":"terminal_sdata","tid":terminal['tid']}}]]
-            );
         }
     }
 }